# Change checksum backend to h5py to improve CI performance