--- a/tests/checksum/backend/openpmd_backend.py
+++ b/tests/checksum/backend/openpmd_backend.py
@@ -17,8 +17,7 @@ class Backend:
         ''' Constructor: store the dataset object
         '''
 
-        # test
-        self.dataset = OpenPMDTimeSeries(filename)
+        self.dataset = OpenPMDTimeSeries(filename, backend='h5py')
 
     def fields_list(self):
         ''' Return the list of fields defined on the grid