@AlexanderSinn 2022-03-28
Does this fix #717?