@MaxThevenet 2022-03-31
@AlexanderSinn thanks for this PR! Yes it seems to address #717, although I originally expected that CI was slowed down because HiPACE++ was slower, not because the Python analysis scripts were. Could you merge dev in there?
@RemiLehe @ax3l do you see a reason why the post-processing should be slower with the openPMD-api backend than with the hdf5 backend of openPMD-viewer? In case that helps, here are the versions used for CI
```
Collecting openpmd-viewer
[297](https://github.com/Hi-PACE/hipace/runs/5718094870?check_suite_focus=true#step:3:297)
  Downloading openPMD_viewer-1.3.0-py3-none-any.whl (85 kB)
[298](https://github.com/Hi-PACE/hipace/runs/5718094870?check_suite_focus=true#step:3:298)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 85.2/85.2 KB 24.0 MB/s eta 0:00:00
[299](https://github.com/Hi-PACE/hipace/runs/5718094870?check_suite_focus=true#step:3:299)
Collecting openpmd-api
[300](https://github.com/Hi-PACE/hipace/runs/5718094870?check_suite_focus=true#step:3:300)
  Downloading openPMD_api-0.14.4-cp38-cp38-manylinux_2_12_x86_64.manylinux2010_x86_64.whl (6.9 MB)

Successfully installed cycler-0.11.0 h5py-3.6.0 kiwisolver-1.4.1 matplotlib-3.2.2 numpy-1.22.3 openpmd-api-0.14.4 openpmd-viewer-1.3.0 python-dateutil-2.8.2 scipy-1.8.0 tqdm-4.63.1
```