@SeverinDiederichs 2022-03-31
The issue was identified to be caused by the upgrade of openPMD viewer from 1.2 to 1.3:

The last fast CI was on PR #687 and the first slow CI was on PR #689 was slow.

Fast CI was with openPMD viewer 1.2 (from the log of #687):
```
Collecting openpmd-viewer
  Downloading openPMD_viewer-1.2.0-py3-none-any.whl (84 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 84.5/84.5 KB 26.7 MB/s eta 0:00:00
Collecting openpmd-api
  Downloading openPMD_api-0.14.4-cp38-cp38-manylinux_2_12_x86_64.manylinux2010_x86_64.whl (6.9 MB)
```

slow CI was with openPMD viewer 1.3 (from the log of #689)
```
Collecting openpmd-viewer
  Downloading openPMD_viewer-1.3.0-py3-none-any.whl (85 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 85.2/85.2 KB 26.3 MB/s eta 0:00:00
Collecting openpmd-api
  Downloading openPMD_api-0.14.4-cp38-cp38-manylinux_2_12_x86_64.manylinux2010_x86_64.whl (6.9 MB)
```