@RemiLehe 2022-03-31
Thanks for reporting this. 
I wonder if this is related of this PR:
https://github.com/openPMD/openPMD-viewer/pull/332/files
It changes the way in which we read the fields. @SeverinDiederichs Are you able to tell whether the slow-down is related to how we read fields as opposed to particles?