@SeverinDiederichs 2022-03-31
This will require some more testing, because all the tests that take significantly longer read out both fields and particles.