@AlexanderSinn 2022-03-31
https://github.com/openPMD/openPMD-viewer/pull/332/files seems to indeed introduce the Issue, however my testing suggests that the slow-down actually occurs when reading in particles with the get_data() function.