@ax3l 2022-04-05
Sorry for being so slow on GitHub pings, feel free to also ping on Slack as well when I overlook something important :(

Thank you for triaging & finding the root cause of this is from https://github.com/openPMD/openPMD-viewer/pull/332 / https://github.com/openPMD/openPMD-viewer/pull/334.

You used HDF5 files, right? (Not json, which is for testing & very slow, ofc.)