@AlexanderSinn 2022-04-05
Yes, HDF5 files