@ax3l 2022-04-06
I am amazed that `np.zeros_like` already takes ~25% of the runtime to read this.