@ax3l 2022-04-12
I drafted a potential fix in https://github.com/openPMD/openPMD-viewer/pull/340