@ax3l 2022-04-15
Fixed via `openPMD-viewer 1.4.0` :tada: 
https://github.com/openPMD/openPMD-viewer/releases/tag/1.4.0

Thank you @AlexanderSinn for the help! :)